fix: factor desktop onboarding helpers
