+++ src/test-kernel/desktop/ElectronRendererShell.vitest.mts
@@ -14,6 +14,13 @@ function readRendererMain(): string {
   );
 }
 
+function readRendererOnboarding(): string {
+  return readFileSync(
+    repoPath('packages/desktop/src/renderer/desktopOnboarding.ts'),
+    'utf8',
+  );
+}
+
 describe('desktop renderer shell', () => {
   it('mounts the reused launcher, progress, and settings Lit apps', () => {
     const rendererMain = readRendererMain();
@@ -66,6 +73,7 @@ describe('desktop renderer shell', () => {
 
   it('mounts desktop-only first-run onboarding controls', () => {
     const rendererMain = readRendererMain();
+    const rendererOnboarding = readRendererOnboarding();
 
     expect(rendererMain).toContain('createFirstRunWalkthrough');
     expect(rendererMain).toContain('DESKTOP_ONBOARDING_COMMANDS.REQUEST_STATE');
@@ -74,11 +82,11 @@ describe('desktop renderer shell', () => {
     expect(rendererMain).toContain(
       'canOpen: () => !firstRunWalkthrough.isVisible()',
     );
-    expect(rendererMain).toContain('previousFocus');
-    expect(rendererMain).toContain("document.addEventListener('focusin'");
-    expect(rendererMain).toContain("'keydown'");
-    expect(rendererMain).toContain('isCommandPaletteShortcut(event)');
-    expect(rendererMain).toContain("event.key !== 'Tab'");
+    expect(rendererOnboarding).toContain('previousFocus');
+    expect(rendererOnboarding).toContain("document.addEventListener('focusin'");
+    expect(rendererOnboarding).toContain("'keydown'");
+    expect(rendererOnboarding).toContain('isCommandPaletteShortcut(event)');
+    expect(rendererOnboarding).toContain("event.key !== 'Tab'");
   });
 
   it('mounts an in-app log viewer with copy and export actions', () => {
